% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addins.R
\name{pick_color_addin_mosaic}
\alias{pick_color_addin_mosaic}
\title{Pick Color as Mosaic}
\usage{
pick_color_addin_mosaic()
}
\description{
Run \code{pick_color(type = 'mosaic')} as addin, you may not run
this directly by code, since it has not arguments
}
