% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_color.R
\name{print_color}
\alias{print_color}
\title{Print the colors}
\usage{
print_color(x, type = "mosaic", blank.len = 3)
}
\arguments{
\item{x}{A literal vector, a palette, or a ggplot object}

\item{type}{A character. The style of printing. The default is \code{mosaic}. It should be one of the \code{mosaic},
\code{ribbon}, or \code{m}, \code{r}}

\item{blank.len}{A integer. The length of color mosaic. The default value is 3.}
}
\description{
Print out the colors used by a literal vector,
a palette, or a ggplot object
}
\examples{

print_color(colors(), type = "mosaic", blank.len = 3)
print_color(palette(), type = "ribbon")

}
