% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addins.R
\name{z_pick_color_addin_ribbon}
\alias{z_pick_color_addin_ribbon}
\title{Pick Color as Ribbons}
\usage{
z_pick_color_addin_ribbon()
}
\description{
Run \code{pick_color(type = 'ribbon')} as addin, you may not run
this directly by code, since it comes without arguments
}
